refactor: update API endpoints to use public events
