+++ src/app/(homepage)/sections/events/index.tsx
@@ -45,7 +45,7 @@ function getEventsUrl(year: number, month: number) {
   const formattedEnd = format(endDate, "yyyy-MM-dd");
 
   // 4. Build the URL
-  return `${API_URL}/events/public?from=${formattedStart}&to=${formattedEnd}`;
+  return `${API_URL}/public/events?from=${formattedStart}&to=${formattedEnd}`;
 }
 
 async function fetchEvents(
